requirements.txt
django==3.1
djangorestframework==3.12
drf-yasg==1.21.4
python-dotenv==0.21.0
django-filter==21.1
